# openapi.yaml
openapi: 3.0.3
info:
  title: Document Analyzer RAG API
  version: 2.0.0
  description: Session-based PDF ingestion and RAG chat backend.
servers:
  - url: http://localhost:4000
components:
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: object
...
    ChatResponseSection:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    ChatResponseSchema:
      type: object
      nullable: true
      properties:
        format:
          type: string
          enum: [structured_sections, structured_markdown]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ChatResponseSection'
    ChatResponseData:
      type: object
      properties:
        answer:
          type: string
        message:
          type: string
        formattedAnswer:
          type: string
        responseSchema:
          $ref: '#/components/schemas/ChatResponseSchema'
        responseStyle:
          type: string
          enum: [structured, plain]
        sources:
          type: array
          items:
            type: object
            properties:
              pdfId:
                type: integer
              chunkId:
                type: integer
              score:
                type: number
        usedChunksCount:
          type: integer
        sessionTitle:
          type: string
paths:
...
  /api/v1/sessions/{sessionId}/chat:
    post:
      summary: Run chat query against indexed session documents. 
      description: |
        If no documents exist for the session, the backend immediately returns a helpful assistant message instructing the user to upload a file (HTTP 200) instead of throwing a PDF_NOT_READY error.
        Session title is generated automatically on the 2nd message, and may be refined automatically from the 6th message onwards.
        When `stream=true`, the connection emits heartbeat pings, includes `retry: 3000` for reconnects, and strictly caps generation at 120s.
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
        - in: query
          name: stream
          schema:
            type: boolean
          description: Enable SSE token streaming
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      text:
                        type: string
                responseStyle:
                  type: string
                  enum: [structured, plain]
                  description: Response formatting style. Defaults to structured.
      responses:
        '200':
          description: Chat answer or SSE stream response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChatResponseData'
            text/event-stream:
              schema:
                type: string
        '202':
          description: Async job accepted
...